+++ crates/mojito-checker/src/checker/scopes.rs
@@ -99,6 +99,7 @@ impl Checker {
         self.aggregate_origin_scopes.push(HashMap::new());
         self.aggregate_field_origin_scopes.push(HashMap::new());
         self.storage_origin_demand_scopes.push(HashMap::new());
+        self.immutable_origin_binder_scopes.push(HashMap::new());
         self.reference_parameter_scopes.push(HashMap::new());
         self.callable_origin_scopes.push(HashMap::new());
     }
@@ -116,6 +117,7 @@ impl Checker {
         self.aggregate_origin_scopes.pop();
         self.aggregate_field_origin_scopes.pop();
         self.storage_origin_demand_scopes.pop();
+        self.immutable_origin_binder_scopes.pop();
         self.reference_parameter_scopes.pop();
         self.callable_origin_scopes.pop();
     }